Url Detection and No test display
If page name is empty, "index" will be set at default.  This will allow
surveys to be on invisible index pages.   Survey must have "index" as
page name to show.

Also, If no test, instead of warning message, shows 1 em high Surveyor
elite logo.  This allows a site to have the JS file in the head and the
initializing div on every page without distractions.